+++ src/modules/export/index.ts
@@ -2,4 +2,5 @@ export { createPdfOperationEngine } from "./pdfOperationEngine";
 export type { PdfOperationEngine } from "./pdfOperationEngine";
 export { createMemoryPdfExportStorage, createPdfExportService } from "./pdfExportService";
 export type { PdfExportService, PdfExportStorage } from "./pdfExportService";
-
+export { createPdfOutputToolsExportRequest, suggestPdfOutputToolsPath } from "./outputTools";
+export type { CreatePdfOutputToolsExportRequestInput } from "./outputTools";